+++ ts/packages/spl-token/src/program.ts
@@ -14,15 +14,14 @@ interface GetProgramParams {
 
 export function splTokenProgram(params?: GetProgramParams): Program<SplToken> {
   return new Program<SplToken>(
-    IDL,
-    params?.programId ?? SPL_TOKEN_PROGRAM_ID,
+    params?.programId ? { ...IDL, address: params.programId.toString() } : IDL,
     params?.provider,
     new SplTokenCoder(IDL)
   );
 }
 
 type SplToken = {
-  address: "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA";
+  address: string;
   metadata: {
     name: "splToken";
     version: "3.3.0";